Added support for more fancy logarithmic plotting as well as Savitzky-Golay smoothing
The style system has been reworked to allow parameterisation of styles,
enabling more configurability.

The savgol and fancylog styles were added.
The savgol style enables Savitzky-Golay smoothing, and is configurable with
two parameters:
	samples: number of samples to use for polynomial fitting
	order  : polynomial order to fit
The fancylog style enables the generation of fancier-looking logarithmic
plots, and is configurable with two parameters:
	cmin: lowest number to display
	cmax: highest number to display
The fancylog style may be incompatible with Fiddle, but this has yet to
be tested.

Finally, all leading spaces in tessierPlot.py were changed to tabs, and
the styles constant in tessierStyles.py was renamed to STYLE_FUNCS to be
more distinguishable from the new STYLE_SPECS constant and to follow
coding conventions (capitalised constants) better.