Added Python3 support
Changed all "print" commands to "print()" function calls, all instances
of "Exception, e" to "Exception as e" and all format strings to the new
format ({:s} instead of %s etc. and using the "".format() function for
applying the formatting instead of "" % ()). Because of this,
TessierPlot can now run in both Python2 and Python3, whereas before it
could only run in Python2.