More support for opening gzip files
Changed tessierplot so the canopener module is no longer used. This would give crc errors on windows
Changes tessierview to only plot the first found file matching the filemask. This is handy on the measurement computer so it will only plot either .dat or .dat.gz file and not double plot the same data.