make unzipped .dat files be plotted again and add some code to make .dat files with different header structure work
- in data.py: for unzipped dat files, the parser got stuck due to
'filebuffer' being 'None'. Fixed that by adding the filebuffer when
calling the parser constructor.
- in data.py: added a second regex for the .dat file header for data
files without the 'start' and 'end' line in the header
in plot.py: added a fallback option in case the axis label format is not
as expected by the regex in parseUnitAndNameFromColumnName. then it just
uses the coord_keys as is for the axis label