added pyqt5 compatibility
tessierplot now works with both pyqt4 and pyqt5. pyqt5 is supposed to be
(and indeed seems) less prone to crash randomly. So if you can, update
to pyqt5. Maybe we can get rid of pyqt4 compatibility in the future
then. pyqt4 and pyqt5 are not compatible, it's a hell hole to work with
both.